commit
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/locate-candidates-to-overflow@45489 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/core/algorithm/FaultLocalizationByImport.java
@@ -197,7 +197,7 @@ public class FaultLocalizationByImport implements Algorithm {
     } catch (IOException e) {
       throw new CPAException("Could not deserialize faults", e);
     }
-    logger.logf(Level.INFO, "Finished parsing %d faults.", faults.size());
+    logger.log(Level.INFO, "Finished parsing %d faults.", faults.size());
     AlgorithmStatus status;
     if (algorithmActivated) {
       logger.log(Level.INFO, "Run configured analysis...");
